requirements.txt
# Core ML Libraries (pinned versions for reproducibility)
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0

# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.0  # Required for FastAPI TestClient

# MLflow for experiment tracking
mlflow==2.8.1

# Data processing and visualization
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# Utilities
tqdm==4.66.1
